+++ src/induction/traits/ConditionalProbabilityDistribution.scala
@@ -4,9 +4,9 @@ trait ConditionalProbabilityDistribution {
 	
   val variableVocabSize: Int
   val givenVariableVocabSize: Int
-  val probabilities : Array[Double]
+  val probabilities : Array[BigDecimal]
   
-  def apply(givenAnotherVariable:Int, variable:Int) : Double = {
+  def apply(givenAnotherVariable:Int, variable:Int) : BigDecimal = {
     return probabilities(getIndexFor(variable,givenAnotherVariable))
   }
   
@@ -21,7 +21,7 @@ trait ConditionalProbabilityDistribution {
     
     for (g <- 0 until givenVariableVocabSize) {
     	for (v <- 0 until variableVocabSize) {
-    		val p : Double = this.apply(v,g)
+    		val p : BigDecimal = this.apply(v,g)
     		s.append(toString(v,g,p,variableVocab,givenVariableVocab))
     		s.append('\n')
     	} 
@@ -30,7 +30,7 @@ trait ConditionalProbabilityDistribution {
     return s.toString
   }
   
-  def toString(variable:Int, givenAnotherVariable:Int, probability:Double, variableVocab:Option[Vocabulary]=None, givenVariableVocab:Option[Vocabulary]=None) : String = {
+  def toString(variable:Int, givenAnotherVariable:Int, probability:BigDecimal, variableVocab:Option[Vocabulary]=None, givenVariableVocab:Option[Vocabulary]=None) : String = {
     val variableString = variableVocab match {
       case None => variable.toString
       case Some(vocab) => vocab.getString(variable)
